test(github-client): rate-limit統合テストを別ファイルへ分離
【背景】test_github_client.py が120件のテストを抱え、lifecycle・request・
rate-limit・ETagの責務が単一ファイルに混在して可読性を損なっていた。

【変更内容】rate-limit関連27テストをAST同一のまま
test_github_client_rate_limit.py へ移動し、純粋関数単体テスト
(test_github_rate_limit.py) との責務境界をmodule docstringに明記した。
分割元に残った空行の整形崩れも併せて修正している。

品質ゲート: 1443 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors
